Git improve added to repository

+++ lib/utils.py
@@ -72,10 +72,13 @@ def clean_files(files):
     items = files if hasattr(files, '__iter__') else [files]
     for item in items:
         fname = item[0] if hasattr(item, '__iter__') else item
-        if os.path.isfile(fname):
-            os.remove(fname)
-        elif os.path.isdir(fname):
-            shutil.rmtree(fname)
+        if fname != "/":
+            if os.path.isfile(fname):
+                os.remove(fname)
+            elif os.path.isdir(fname):
+                shutil.rmtree(fname)
+        else:
+            logger.error("Invalid target path: '/'. Are you trying to delete your root path?")
 
 
 def simplify_path(b_info):
@@ -88,19 +91,24 @@ def simplify_path(b_info):
         List of dictionaries with branches' info
     """
     logger.debug("Deleting all common branches' path")
-    repeated = True
-    while repeated:
-        piece_path = []
-        for branch in b_info:
-            piece_path.append(branch['path'].split('/', 1)[0])
-        word = piece_path[0]
+    if len(b_info) > 1:
         repeated = True
-        for each in piece_path:
-            if each != word:
-                repeated = False
-        if repeated:
+        while repeated:
+            piece_path = []
             for branch in b_info:
-                branch.update({'path': branch['path'].split('/', 1)[1]})
+                piece_path.append(branch['path'].split('/', 1)[0])
+            word = piece_path[0]
+            repeated = True
+            for each in piece_path:
+                if each != word:
+                    repeated = False
+            if repeated:
+                for branch in b_info:
+                    branch.update({'path': branch['path'].split('/', 1)[1]})
+    else:
+        if len(b_info) > 0:
+            branch = b_info[0]
+            branch.update({'path': os.path.basename(branch['path'])})
     logger.debug("Common paths deleted")
     return b_info
 
@@ -164,7 +172,7 @@ def decompress_files(name, dest_folder):
     base_folder = None
     for fname in name_list:
         if os.path.basename(fname.name) == 'database_dump.b64' or \
-            os.path.basename(fname.name) == 'database_dump.sql':
+           os.path.basename(fname.name) == 'database_dump.sql':
             base_folder = os.path.dirname(fname.name)
             break
 
@@ -197,6 +205,7 @@ def dump_database(dest_folder, database_name, super_user_pass, host, port):
     with open(dump_name, "w") as fout:
         fout.write(binary_data)
     return dump_name
+
 
 def generate_backup_name(database_name, reason=False):
     """Generates the backup name accordint to the following standar:
@@ -332,6 +341,7 @@ def test_connection(db_name, host=False, port=8069, user=False,
                         "instance properly with the supplied password", user)
     return True
 
+
 def decode_b64_file(src, dst):
     """ Read src base64 encoded file and output its conntent to dst file
     """
@@ -339,6 +349,7 @@ def decode_b64_file(src, dst):
         with open(dst, 'w') as destination_file:
             for line in source_file:
                 destination_file.write(base64.b64decode(line))
+
 
 def restore_direct(backup, odoo_config, working_dir, container_name=None):
     """ Restore a pg_dump in sql format or b64 generated with the odoo webservice
@@ -392,18 +403,19 @@ def pase_odoo_configfile(filename):
         logger.error('configuration file "%s" not found', filename)
         return None
     if config.get('options', 'db_host') != 'False':
-        res.update({'db_host' : config.get('options', 'db_host')})
+        res.update({'db_host': config.get('options', 'db_host')})
     else:
-        res.update({'db_host' : 'localhost'})
+        res.update({'db_host': 'localhost'})
     if config.get('options', 'db_port') != 'False':
-        res.update({'db_port' : config.get('options', 'db_port')})
+        res.update({'db_port': config.get('options', 'db_port')})
     else:
-        res.update({'db_port' : 5432})
-    res.update({'db_user' : config.get('options', 'db_user')})
-    res.update({'db_password' : config.get('options', 'db_password')})
-    res.update({'data_dir' : config.get('options', 'data_dir')})
-    
+        res.update({'db_port': 5432})
+    res.update({'db_user': config.get('options', 'db_user')})
+    res.update({'db_password': config.get('options', 'db_password')})
+    res.update({'data_dir': config.get('options', 'data_dir')})
+
     return res
+
 
 def pgdump_database(dest_folder, database_config):
     """ Dumps database using pg_dump in sql format
@@ -429,6 +441,7 @@ def pgdump_database(dest_folder, database_config):
             logger.error('Could not dump database, error message: %s', error.stderr_output)
         return None
     return dump_name
+
 
 def pgrestore_database(dump_name, database_config):
     """ Restores a databse dump in sql plain format, tries to create database if not exists
@@ -462,6 +475,7 @@ def pgrestore_database(dump_name, database_config):
         return None
     return True
 
+
 def get_docker_env(container_name, docker_url="unix://var/run/docker.sock"):
     """ Get env vars from a docker container
     Args:
@@ -486,6 +500,7 @@ def get_docker_env(container_name, docker_url="unix://var/run/docker.sock"):
         res.update({name: value})
 
     return res
+
 
 def restore_docker_filestore(src_folder, database_name, container_name,
                              docker_url="unix://var/run/docker.sock"):
@@ -532,6 +547,7 @@ def restore_docker_filestore(src_folder, database_name, container_name,
     cli.execute(container_name, "mv /tmp/filestore {}".format(fs_name))
     cli.execute(container_name, "chown -R {0}:{0} {1}".format(env_vars.get('ODOO_USER'), fs_name))
 
+
 def restore_instance_filestore(src_folder, odoo_config):
     """ Restore filestore to a instance directly
     Args:
@@ -544,7 +560,9 @@ def restore_instance_filestore(src_folder, odoo_config):
                                odoo_config.get('database'))
     shutil.move(src_folder, dest_folder)
 
-def backup_database_direct(odoo_config, dest_folder, reason=False, tmp_dir=False):
+
+def backup_database_direct(odoo_config, dest_folder, reason=False,
+                           tmp_dir=False):
     """ Receive database name and back it up
 
     Args:
@@ -594,15 +612,14 @@ def parse_docker_config(container_name, docker_url="unix://var/run/docker.sock")
 
     env_vars = get_docker_env(container_name)
     res = {
-        'db_host' : env_vars.get('DB_HOST', None),
-        'db_port' : env_vars.get('DB_PORT', 5432),
-        'db_user' : env_vars.get('DB_USER', 'odoo'),
-        'db_password' : env_vars.get('DB_PASSWORD'),
+        'db_host': env_vars.get('DB_HOST', None),
+        'db_port': env_vars.get('DB_PORT', 5432),
+        'db_user': env_vars.get('DB_USER', 'odoo'),
+        'db_password': env_vars.get('DB_PASSWORD'),
     }
 
     if '172.17.42.1' == res.get('db_host'):
         res.update({'db_host': '127.0.0.1'})
-
 
     cli = Client(base_url=docker_url)
     try:
@@ -613,7 +630,7 @@ def parse_docker_config(container_name, docker_url="unix://var/run/docker.sock")
         else:
             logger.error(error.explanation)
         return None
-                
+
     volumes = inspected.get('Volumes')
     for mnt, volume in volumes.iteritems():
         if '.local/share/Odoo' in mnt and volume:
@@ -627,6 +644,7 @@ def parse_docker_config(container_name, docker_url="unix://var/run/docker.sock")
                       'wont be able to backup attachments'))
     return res
 
+
 def dropdb_direct(database_config):
     """ Drop a database using the config provided
     Args:
@@ -637,7 +655,7 @@ def dropdb_direct(database_config):
     if database_config.get('db_password') != 'False':
         os.environ['PGPASSWORD'] = database_config.get('db_password')
     dropdb_cmd = 'dropdb {database} -U {db_user} -p {db_port} -h {db_host}' \
-    .format(**database_config)
+                 .format(**database_config)
     shell = spur.LocalShell()
     try:
         shell.run(shlex.split(dropdb_cmd))
@@ -646,6 +664,7 @@ def dropdb_direct(database_config):
         return None
     else:
         return True
+
 
 def remove_attachments(odoo_config, container=False):
     """Remove attachements dolder
